updating stops

+++ src/Dijitle.Metra.Data/Stops.cs
@@ -9,8 +9,8 @@ namespace Dijitle.Metra.Data
         public string stop_id {  get; private set; }
         public string stop_name { get; private set; }
         public string stop_desc { get; private set; }
-        public decimal stop_lat { get; private set; }
-        public decimal stop_lon { get; private set; }
+        public double stop_lat { get; private set; }
+        public double stop_lon { get; private set; }
         public string zone_id { get; private set; }
         public string stop_url { get; private set; }
         public bool wheelchair_boarding { get; private set; }
@@ -22,8 +22,8 @@ namespace Dijitle.Metra.Data
             stop_id = csv[0].Trim();
             stop_name = csv[1].Trim();
             stop_desc = csv[2].Trim();
-            stop_lat = Convert.ToDecimal(csv[3].Trim());
-            stop_lon = Convert.ToDecimal(csv[4].Trim());
+            stop_lat = Convert.ToDouble(csv[3].Trim());
+            stop_lon = Convert.ToDouble(csv[4].Trim());
             zone_id = csv[5].Trim();
             stop_url = csv[6].Trim();
             wheelchair_boarding = (csv[7].Trim() == "1");
